moves conversion fns to service/types/
