+++ rust/personhog-replica/src/service/types/mod.rs
@@ -1,0 +1,2 @@
+mod group;
+mod person;
